Extrair progresso da CLI para modulo dedicado
