+++ src/ayvu/cli.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-from dataclasses import dataclass
 from pathlib import Path
 from typing import Optional
 
@@ -10,6 +9,7 @@ from rich.progress import BarColumn, Progress, SpinnerColumn, TaskProgressColumn
 from rich.table import Table
 
 from .cache import TranslationCache
+from .cli_progress import TranslationProgress
 from .domain import LanguagePair, OutputPlan, TranslationOptions
 from .epub_io import TranslationReport, extract_markdown, inspect_epub, translate_epub
 from .preflight import PreflightError, run_translation_preflight
@@ -19,74 +19,6 @@ from .validation import validate_output_epub
 
 app = typer.Typer(help="Translate local EPUB files with a local HTTP translator.")
 console = Console()
-
-
-@dataclass
-class TextProgressCounters:
-    translated: int = 0
-    cache: int = 0
-    dry_run: int = 0
-    error: int = 0
-
-    def record(self, status: str) -> None:
-        if status == "translated":
-            self.translated += 1
-            return
-        if status == "cache":
-            self.cache += 1
-            return
-        if status == "dry_run":
-            self.dry_run += 1
-            return
-        if status == "error":
-            self.error += 1
-            return
-        raise ValueError(f"Unknown text progress status: {status}")
-
-    @property
-    def processed(self) -> int:
-        return self.translated + self.cache + self.dry_run + self.error
-
-    def new_count(self, dry_run: bool) -> int:
-        if dry_run:
-            return self.dry_run
-        return self.translated
-
-
-class TranslationProgress:
-    def __init__(self, progress: Progress, dry_run: bool) -> None:
-        self._progress = progress
-        self._dry_run = dry_run
-        self._counters = TextProgressCounters()
-        self._chapter_task = progress.add_task("Chapters", total=None)
-        self._text_task = progress.add_task("Texts", total=None)
-
-    def chapter_started(self, index: int, total: int, name: str) -> None:
-        self._progress.update(
-            self._chapter_task,
-            total=total,
-            description=self._chapter_description(index, total, name),
-        )
-
-    def chapter_done(self, index: int, total: int, name: str, _stats: object) -> None:
-        self._progress.advance(self._chapter_task)
-        self._progress.update(self._chapter_task, description=self._chapter_description(index, total, name))
-
-    def text_processed(self, status: str) -> None:
-        self._counters.record(status)
-        self._progress.advance(self._text_task)
-        self._progress.update(self._text_task, description=self._text_description())
-
-    def _chapter_description(self, index: int, total: int, name: str) -> str:
-        return f"Chapters {index}/{total}: {_shorten(name)}"
-
-    def _text_description(self) -> str:
-        new_label = "would translate" if self._dry_run else "new"
-        new_count = self._counters.new_count(self._dry_run)
-        return (
-            f"Texts {self._counters.processed} | {new_label} {new_count} | "
-            f"cache {self._counters.cache} | errors {self._counters.error}"
-        )
 
 
 @app.command()
@@ -336,12 +268,6 @@ def _confirm_existing_output_overwrite(output_path: Path) -> bool:
     console.print(f"[yellow]Output path:[/yellow] {output_path}")
     console.print("[yellow]Translated EPUB already exists.[/yellow]")
     return typer.confirm("Overwrite existing translated EPUB?", default=False)
-
-
-def _shorten(text: str, max_length: int = 50) -> str:
-    if len(text) <= max_length:
-        return text
-    return text[: max_length - 3] + "..."
 
 
 if __name__ == "__main__":
